refactor: extract CacheableStats trait and add comprehensive PHPDoc
Major refactoring and documentation improvements:

- Create CacheableStats trait to reduce code duplication in Dashboard.php
  - 14% code reduction (1,015 → 875 lines)
  - Centralized cache handling with fallback support
  - Methods: cacheOrFallback, cachedStats, executeOrDefault, forgetCacheKeys

- Add comprehensive PHPDoc to 40+ Livewire components
  - Class-level documentation with @property annotations
  - Method-level documentation with @param and @return types
  - Improved IDE support and type inference

- Create test suites for new features:
  - HelpContentServiceTest for inline help system
  - DocsControllerTest for documentation controller
  - Factory classes for HelpContent models

- Extract DevFlowSelfManagement sub-components:
  - CacheManager for cache operations
  - DeploymentActions for deploy/maintenance actions
  - LogViewer for log viewing functionality

+++ app/Models/HelpContentRelated.php
@@ -18,7 +18,7 @@ class HelpContentRelated extends Model
     ];
 
     protected $casts = [
-        'relevance_score' => 'integer',
+        'relevance_score' => 'float',
     ];
 
     public function helpContent(): BelongsTo
